refactor(sqs): consolidate duplicated FakeAmazonSQS test fake
The two FakeAmazonSQS fakes (cristin-import and s3-import-commons) were
byte-for-byte identical apart from their package. Move a single canonical
copy into s3-import-commons test fixtures (next to SqsBatchMessenger, the
class it supports) and consume it from cristin-import via
testImplementation(testFixtures(project(":s3-import-commons"))). cristin-import
already depends on s3-import-commons, so this adds no new module coupling.

Declare aws.sdk2.sqs as testFixturesApi (the fake exposes SqsClient on its API
surface) as flagged by the dependency-analysis build health check, and drop
the dead getQueueUrls()/queueUrls tracking that no test ever read.
