+++ cristin-import/src/test/java/no/unit/nva/cristin/lambda/CristinRerunEventEmitterTest.java
@@ -12,8 +12,8 @@ import java.net.URI;
 import java.util.List;
 import java.util.stream.Collectors;
 import no.unit.nva.commons.json.JsonUtils;
-import no.unit.nva.cristin.FakeAmazonSQS;
 import no.unit.nva.events.models.EventReference;
+import no.unit.nva.publication.s3imports.utils.FakeAmazonSQS;
 import no.unit.nva.s3.S3Driver;
 import no.unit.nva.stubs.FakeContext;
 import no.unit.nva.stubs.FakeS3Client;
