+++ s3-import-commons/src/testFixtures/java/no/unit/nva/publication/s3imports/utils/FakeAmazonSQS.java
@@ -1,8 +1,7 @@
-package no.unit.nva.cristin;
+package no.unit.nva.publication.s3imports.utils;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 import software.amazon.awssdk.services.sqs.SqsClient;
 import software.amazon.awssdk.services.sqs.model.SendMessageBatchRequest;
 import software.amazon.awssdk.services.sqs.model.SendMessageBatchRequestEntry;
@@ -12,11 +11,6 @@ import software.amazon.awssdk.services.sqs.model.SendMessageBatchResultEntry;
 public class FakeAmazonSQS implements SqsClient {
 
   private final List<String> messageBodies = new ArrayList<>();
-  private final List<String> queueUrls = new ArrayList<>();
-
-  public List<String> getQueueUrls() {
-    return queueUrls;
-  }
 
   public List<String> getMessageBodies() {
     return messageBodies;
@@ -38,11 +32,8 @@ public class FakeAmazonSQS implements SqsClient {
     sendMessageBatchRequest.entries().stream()
         .map(SendMessageBatchRequestEntry::messageBody)
         .forEach(messageBodies::add);
-    queueUrls.add(sendMessageBatchRequest.queueUrl());
     var messageBatchResultEntries =
-        sendMessageBatchRequest.entries().stream()
-            .map(this::convertToResult)
-            .collect(Collectors.toList());
+        sendMessageBatchRequest.entries().stream().map(this::convertToResult).toList();
     return SendMessageBatchResponse.builder().successful(messageBatchResultEntries).build();
   }
